+++ crypto-scraper/app/Console/Commands/Bitcointalk/Parsers/UserProfile.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 namespace App\Console\Commands\Bitcointalk\Parsers;
 
 use App\Console\Base\Bitcointalk\BitcointalkParser;
-use App\Console\Base\CryptoCurrency;
+use App\Console\Constants\CryptoCurrency;
 use App\Models\ParsedAddress;
 use App\Models\Pg\Category;
