+++ crypto-scraper/app/Models/ParsedAddress.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace App\Models;
 
-use App\Console\Base\Utils;
+use App\Console\Base\Common\Utils;
 
 class ParsedAddress {
     private $owner, $url, $label, $source, $address, $type, $explicitCategory;
