+++ crypto-scraper/tests/Unit/CreateTSVDataTest.php
@@ -2,7 +2,7 @@
 
 namespace Tests\Unit;
 
-use App\Console\Base\Utils;
+use App\Console\Base\Common\Utils;
 use Tests\TestCase;
 
 class CreateTSVDataTest extends TestCase
